Add LOG_CHARSET to declare the output charset
Allow folks to declare a charset for the rendered pages, because the
browser's fallback charset may not match the encoding used in the
channels.  Ideally the incoming data would also be decoded on recipt
and then re-encoded to LOG_CHARSET before saving, but since all my
channels use UTF-8, I don't have to worry about that.

Docs for:
* HTTP's charset parameter [1].
* HTML5's meta charset [2,3].

The MDN page also points out [3]:

> This <meta> element must be inside the <head> element and within the
> 512 first bytes of the page, as some browsers only look at these
> first bytes before choosing a character set for the page.

which is why I've made it the first entry in <head>.

[1]: http://www.w3.org/International/O-HTTP-charset
[2]: http://www.w3.org/TR/html-markup/meta.charset.html
[3]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta